forecast: share weight normalization
