+++ src/mtdata/forecast/common.py
@@ -41,6 +41,29 @@ def log_returns_from_prices(prices: np.ndarray, eps: float = 1e-12) -> np.ndarra
     return np.asarray(rets, dtype=float)
 
 
+def _normalize_weights(weights: Any, size: int) -> Optional[np.ndarray]:
+    if weights is None:
+        return None
+    vals: List[float] = []
+    if isinstance(weights, (list, tuple)):
+        vals = [float(v) for v in list(weights)[:size]]
+    elif isinstance(weights, str):
+        parts = [p.strip() for p in weights.split(",") if p.strip()]
+        vals = [float(p) for p in parts[:size]]
+    else:
+        return None
+    if len(vals) != size:
+        return None
+    arr = np.asarray(vals, dtype=float)
+    if not np.all(np.isfinite(arr)):
+        return None
+    arr = np.clip(arr, a_min=0.0, a_max=None)
+    total = float(np.sum(arr))
+    if total <= 0:
+        return None
+    return arr / total
+
+
 def _extract_forecast_values(Yf: Any, fh: int, method_name: str = "forecast") -> "np.ndarray":
     """Extract forecast values from prediction DataFrame.
