+++ src/mtdata/forecast/forecast_engine.py
@@ -6,7 +6,6 @@ from typing import Any, Dict, Optional, List, Literal, Tuple
 import logging
 import numpy as np
 import pandas as pd
-import math
 
 from ..bootstrap.settings import mt5_config
 from ..shared.constants import TIMEFRAME_MAP, TIMEFRAME_SECONDS
@@ -21,6 +20,7 @@ from ..utils.utils import (
     parse_kv_or_json as _parse_kv_or_json,
 )
 from . import forecast_preprocessing as _forecast_preprocessing
+from .common import _normalize_weights as _normalize_weights_impl
 from .common import (
     default_seasonality,
     fetch_history as _fetch_history,
@@ -73,6 +73,7 @@ _ENSEMBLE_BASE_METHODS = (
 )
 
 logger = logging.getLogger(__name__)
+_normalize_weights = _normalize_weights_impl
 
 
 def _clear_ensemble_dispatch_error() -> None:
@@ -125,29 +126,6 @@ def _append_ensemble_failure(
             if value is not None:
                 payload[str(key)] = value
     failures.append(payload)
-
-
-def _normalize_weights(weights: Any, size: int) -> Optional[np.ndarray]:
-    if weights is None:
-        return None
-    vals: List[float] = []
-    if isinstance(weights, (list, tuple)):
-        vals = [float(v) for v in list(weights)[:size]]
-    elif isinstance(weights, str):
-        parts = [p.strip() for p in weights.split(',') if p.strip()]
-        vals = [float(p) for p in parts[:size]]
-    else:
-        return None
-    if len(vals) != size:
-        return None
-    arr = np.asarray(vals, dtype=float)
-    if not np.all(np.isfinite(arr)):
-        return None
-    arr = np.clip(arr, a_min=0.0, a_max=None)
-    total = float(np.sum(arr))
-    if total <= 0:
-        return None
-    return arr / total
 
 
 def _ensemble_dispatch_method(
@@ -689,7 +667,6 @@ def forecast_engine(
     """
     try:
         ci_values = None
-        ensemble_meta: Dict[str, Any] = {}
         # Coerce CLI string inputs to proper types
         try:
             horizon = int(horizon) if horizon is not None else 12
